pyproject.toml
[tool.poetry]
name = "recsys-mvp"
version = "0.1.0"
description = ""
authors = ["Quy Dinh <quy.dinh3195@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.9"
torch = "^2.4.1"
torchvision = "^0.19.1"
torchaudio = "^2.4.1"
pandas = "^2.2.2"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
datasets = "^2.21.0"
plotly = "^5.24.0"
pydantic = "^2.9.0"
scikit-learn = "^1.5.1"
evidently = "^0.4.37"
mlflow = "^2.16.0"
pytest = "^8.3.3"
onnx = "^1.16.2"
dill = "^0.3.8"
faiss-cpu = "1.7.4"  # For MacOS if raise to 1.8.0 then kernel crashes
lightning = "^2.4.0"
tensorboard = "^2.18.0"
redis = "^5.1.0"
fastapi = "^0.115.0"
bentoml = "^1.3.7"
papermill = "^2.6.0"
numpy = "<2.0"  # Numpy 2.0 causes a lot of compatibility error


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.2.5"
ipywidgets = "^8.1.5"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
